backend/pyproject.toml
[tool.poetry]
name = "border-badge-backend"
version = "0.1.0"
description = "FastAPI backend for Border Badge travel tracking app"
authors = ["Border Badge Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
httpx = "^0.28.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.6.0"
pyjwt = "^2.9.0"
slowapi = "^0.1.9"
jinja2 = "^3.1.0"
pillow = "^11.0.0"
pillow-heif = "^0.21.0"
email-validator = "^2.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
rcssmin = "^1.2.2"
pytest-asyncio = "^0.24.0"
ruff = "^0.8.0"
pytest-cov = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by ruff format)
    "B008",   # function call in argument defaults (FastAPI pattern)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
